requirements.txt
fastapi
uvicorn
python-dotenv
langchain-google-genai
langgraph
pydantic
requests
google-search-results
duckduckgo-search
